# Optimize allocations in `Hash#compare_by_identity`