* If `self` is an empty Hash, then we don't need to allocate the `tmp` Hash (since there's nothing to re-hash).
* If `self` is non-empty, then we don't need to bother with `ar_force_convert_table`. We're going to discard that table and replace it with the new `identtable` st table that we're stealing out of `tmp`.

So in either case, we get one less allocation.

This simple micro-benchmark clocks it as ~1.7x faster than before:

```
Comparison:
 After: {}.compare_by_identity: 13725103.3 i/s
Before: {}.compare_by_identity:  8085211.5 i/s - 1.70x  slower
```

Oddly, the difference is even larger (~1.85x) if you `GC.disable` (though this "same-ish: difference falls within error" conclusion seems pretty wrong...):

```
Comparison:
 After: {}.compare_by_identity:  9306969.8 i/s
Before: {}.compare_by_identity:  5036135.4 i/s - same-ish: difference falls within error
```

<details><summary><code>benchmark.rb</code></summary>

```ruby
#!/usr/bin/ruby

require "benchmark/ips"

# GC.disable

Benchmark.ips do |x|
  x.report("Before: {}.compare_by_identity") do |times|
    i = 0
    while (i += 1) < times
      {}.compare_by_identity
    end
  end
  
  x.compare!
  x.hold!("compare_by_identity.json")
  
  x.report(" After: {}.compare_by_identity") do |times|
    i = 0
    while (i += 1) < times
      {}.compare_by_identity
    end
  end
  
  x.compare!
end
```

</p>
</details>